+++ ci/doc-preview/xls-to-html-preview.html
@@ -32,7 +32,7 @@
     <script src="https://cdn.bootcss.com/babel-core/5.8.35/browser.min.js"></script>
     <script src="https://cdn.bootcss.com/babel-core/5.8.35/browser-polyfill.min.js"></script>
     <script src="./src/base64.js"></script>
-    <script src="./src/sdk-v0.1.1.js"></script>
+    <script src="./src/sdk-v0.2.1.js"></script>
 
     <script type="text/babel">
